+++ src/main/java/wooteco/subway/path/application/PathService.java
@@ -7,9 +7,9 @@ import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.line.application.LineService;
 import wooteco.subway.line.domain.Line;
 import wooteco.subway.member.domain.LoginMember;
-import wooteco.subway.path.domain.Fare;
-import wooteco.subway.path.domain.FareCalculator;
-import wooteco.subway.path.domain.FareDistance;
+import wooteco.subway.path.domain.fare.Fare;
+import wooteco.subway.path.domain.fare.FareCalculator;
+import wooteco.subway.path.domain.fare.distance.FareDistance;
 import wooteco.subway.path.domain.SubwayPath;
 import wooteco.subway.path.dto.PathResponse;
 import wooteco.subway.path.dto.PathResponseAssembler;
