+++ src/main/java/wooteco/subway/path/domain/fare/distance/DefaultFareDistance.java
@@ -1,4 +1,6 @@
-package wooteco.subway.path.domain;
+package wooteco.subway.path.domain.fare.distance;
+
+import wooteco.subway.path.domain.fare.Fare;
 
 public class DefaultFareDistance implements FareDistance {
