+++ src/main/java/wooteco/subway/path/domain/fare/distance/FareDistance.java
@@ -1,4 +1,6 @@
-package wooteco.subway.path.domain;
+package wooteco.subway.path.domain.fare.distance;
+
+import wooteco.subway.path.domain.fare.Fare;
 
 public interface FareDistance {
     int DEFAULT_FARE_DISTANCE = 10;
