+++ src/main/java/wooteco/subway/path/domain/fare/distance/SecondExtraFareDistance.java
@@ -1,4 +1,6 @@
-package wooteco.subway.path.domain;
+package wooteco.subway.path.domain.fare.distance;
+
+import wooteco.subway.path.domain.fare.Fare;
 
 public class SecondExtraFareDistance implements FareDistance {
     private static final double DISTANCE_STANDARD = 8.0;
