+++ src/test/java/wooteco/subway/path/domain/FareCalculatorTest.java
@@ -13,6 +13,8 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 import wooteco.subway.line.domain.Line;
+import wooteco.subway.path.domain.fare.Fare;
+import wooteco.subway.path.domain.fare.FareCalculator;
 
 class FareCalculatorTest {
     private FareCalculator fareCalculator;
