+++ src/test/java/wooteco/subway/path/domain/fare/distance/FareDistanceTest.java
@@ -1,10 +1,12 @@
-package wooteco.subway.path.domain;
+package wooteco.subway.path.domain.fare.distance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import wooteco.subway.path.domain.fare.Fare;
+import wooteco.subway.path.domain.fare.distance.FareDistance;
 
 class FareDistanceTest {
     @DisplayName("거리에 따라 부가되는 추가요금을 계산한다.")
